% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{aggregateSum}
\alias{aggregateSum}
\title{Peptide/protein aggregate function using sum}
\usage{
aggregateSum(wp)
}
\arguments{
\item{wp}{Matrix with columns corresponding to samples and rows corresponding
to peptide entires.}
}
\value{
A numeric vector with aggregated values
}
\description{
This function should be used from within \code{\link{makePeptideTable}} or
\code{\link{makeProteinTable}}. It aggregates values by calculating the sum
for each column of \code{wp}.
}
\examples{
\dontrun{
library(proteusUnlabelled)
data(proteusUnlabelled)
prodat <- makeProteinTable(pepdat.clean, aggregate.fun=aggregateSum)
}
}
