% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{allProteinTable}
\alias{allProteinTable}
\title{All protein table}
\usage{
allProteinTable(res)
}
\arguments{
\item{res}{A data frame with results from limmaDE.}
}
\value{
A rendered table.
}
\description{
Internal function to create a table with all proteins (bottom of the web page).
}
