% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amzn_aapl_fb}
\alias{amzn_aapl_fb}
\title{amzn_aapl_fb data set}
\format{
A data frame with 4 columns and 1798 rows.
}
\source{
Yahoo Finance
}
\usage{
amzn_aapl_fb
}
\description{
A data frame with the close prices for Amazon, Google and Facebook.
}
\keyword{datasets}
