% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{downlierSigma}
\alias{downlierSigma}
\title{Detect downliers}
\usage{
downlierSigma(v, sigma = 5, trim = 0)
}
\arguments{
\item{v}{An input numeric vector}

\item{sigma}{Z-score limit, default value is 5}

\item{trim}{Number of trimmed points on each side for finding Z-score}
}
\value{
Index of the downlier if detected (that is the lowest intensity Z-score is less than sigma), otherwise zero.
}
\description{
Detect a downlier based on Z-score.
}
