% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\docType{data}
\name{evidenceColumns}
\alias{evidenceColumns}
\title{Evidence columns}
\format{An object of class \code{matrix} with 2 rows and 15 columns.}
\usage{
evidenceColumns
}
\description{
Columns to be read from the evidence file. This matrix contains two rows: first with the name used
in the package and second with the original name.
}
\keyword{datasets}
