% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{goodData}
\alias{goodData}
\title{Create a table indicating good data.}
\usage{
goodData(pdat)
}
\arguments{
\item{pdat}{Peptide or protein \code{proteusData} object.}
}
\value{
A data frame with rows corresponding to peptides/proteins and columns
  corresponding to conditions. Logical values indicate if at least one
  replicate is available (TRUE) or if all replicates are missing (FALSE).
}
\description{
Create a table indicating good data.
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)
detect <- goodData(prodat)

}
