% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{intensityStats}
\alias{intensityStats}
\title{Statistics for an intensity table}
\usage{
intensityStats(intab)
}
\arguments{
\item{intab}{Intensity table}
}
\value{
List of data frames with statistics, one per conditions
}
\description{
Calculate mean, standard deviation, ... for an intensity table (that is peptide or protein,
for all conditions)
}
