% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{jaccardSimilarity}
\alias{jaccardSimilarity}
\title{Jaccard similarity}
\usage{
jaccardSimilarity(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
Jaccard similarity between x and y
}
\description{
Computes Jaccard similarity between "detections" in two vectors of the same
length. A detection is a value, as opposed to NA.
}
\examples{
library(proteusLabelFree)
data(proteusLabelFree)
sim <- jaccardSimilarity(pepdat$tab[,1], pepdat$tab[,2])

}
