% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{jitterPlot}
\alias{jitterPlot}
\title{Jitter plot}
\usage{
jitterPlot(tab, input, pdat, max_points)
}
\arguments{
\item{tab}{Table used to create a plot}

\item{input}{Input variable from shiny server}

\item{pdat}{\code{proteusData} object with data and annotations}

\item{max_points}{Maximum number of points to select}
}
\value{
A ggplot object.
}
\description{
Internal function to create a jitter plot with replicate intensities versus
condition.
}
