% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{limmaDE}
\alias{limmaDE}
\title{Simple differential expression with limma}
\usage{
limmaDE(pdat, formula = "~condition", conditions = NULL,
  transform.fun = log10, sig.level = 0.05)
}
\arguments{
\item{pdat}{Protein \code{proteusData} object.}

\item{formula}{A string with a formula for building the linear model.}

\item{conditions}{A character vector with two conditions for differential
expression. Can be omitted if there are only two condition in \code{pdat}.}

\item{transform.fun}{A function to transform data before differential
expression.}

\item{sig.level}{Significance level for rejecting the null hypothesis.}
}
\value{
A data frame with DE results. "logFC" colum is a log10-fold-change.
}
\description{
\code{limmaDE} is a wrapper around \code{\link{limma}} to perform a
differential expression between a pair of conditions.
}
\details{
Before \code{limma} is called, intensity data are transformed using the
\code{transform.fun} function. The default for this transformation is
\code{log10}. Therefore, by default, the column "logFC" in the output data
frame contains log10 fold change. If you need log2-based fold change, you can
use \code{transform.fun=log2}.

\code{limmaDE} is only a simple wrapper around \code{\link{limma}}, to
perform differential expression between two conditions. For more complicated
designs we recommend using \code{\link{limma}} functions directly.
}
\examples{
res <- limmaDE(xpprodat)

}
