% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{limmaRatioDE}
\alias{limmaRatioDE}
\title{Simple one-sample differential expression with limma}
\usage{
limmaRatioDE(pdat, condition = NULL, transform.fun = log2,
  sig.level = 0.05)
}
\arguments{
\item{pdat}{Protein \code{proteusData} object.}

\item{condition}{Condition name.}

\item{transform.fun}{A function to transform data before differential
expression.}

\item{sig.level}{Significance level for rejecting the null hypothesis.}
}
\value{
A data frame with DE results. "logFC" colum is a log-fold-change
  (using the \code{transform.fun}). Attributes contain
  additional information about the transformation function, significance
  level and condition name.
}
\description{
\code{limmaDE} is a wrapper around \code{\link{limma}} to perform a
differential expression for one condition against the null hypothesis of one.
This is designed to be used with SILAC ratios.
}
\details{
Performs differential expression for one condition, assuming the measured
values are ratios and the null hypothesis is the ratio of 1. The analysis is
done in logarithmic space, so the actual null hypothesis tested is log ratio
= 0. The log fold change returned is the log mean ratio of the condition.

Before \code{limma} is called, intensity data are transformed using the
\code{transform.fun} function. The default for this transformation is
\code{log2}. Therefore, by default, the column "logFC" in the output data
frame contains log2 fold change. If you need log10-based fold change, you can
use \code{transform.fun=log10}.

\code{limmaDE} is only a simple wrapper around \code{\link{limma}}, to
perform differential expression between two conditions. For more complicated
designs we recommend using \code{\link{limma}} functions directly.
}
\examples{
\dontrun{
library(proteusSILAC)
data(proteusSILAC)
prodat.norm <- normalizeData(prodat)
res <- limmaRatioDE(prodat.norm, condition="T48")
}
}
