% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{limmaTable}
\alias{limmaTable}
\title{Create differential expression result.}
\usage{
limmaTable(pdat, ebay, column = "condition")
}
\arguments{
\item{pdat}{Protein intensity structure.}

\item{ebay}{Output from "limmaDE"}

\item{column}{Which column should be used to extract data. The default value is "condition".}
}
\value{
A data frame with DE results.
}
\description{
Use limmaDE output to create a table with DE results.
}
