% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{makeProtein}
\alias{makeProtein}
\title{Helper function, not exported}
\usage{
makeProtein(wp, method, hifly = 3)
}
\arguments{
\item{wp}{Intensity table with selection of peptides for this protein}

\item{method}{Method of creating proteins}

\item{hifly}{Number of high-fliers}
}
\value{
A data frame row with aggregated protein intensities
}
\description{
Make protein from a selection of peptides
}
