% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{makeProteins}
\alias{makeProteins}
\title{Make proteins}
\usage{
makeProteins(evi, peptab, norm = "median", min.peptides = 1)
}
\arguments{
\item{evi}{Evidence table, used only to cross-reference peptides with proteins.}

\item{peptab}{Peptide intensity table.}

\item{norm}{Normalization to use on peptides before converting into proteins.}

\item{min.peptides}{Minimum number of peptides per protein.}

\item{hifly}{The number of top peptides (high-flyers) to be used for protein intensity.}
}
\value{
Protein intensity table.
}
\description{
Create a protein table from the peptide table using high-flyers.
}
