% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{makeProteinTable}
\alias{makeProteinTable}
\title{Make protein table}
\usage{
makeProteinTable(pepdat, aggregate.fun = aggregateHifly, ...,
  min.peptides = 1, ncores = 4)
}
\arguments{
\item{pepdat}{A \code{proteusData} object containing peptide data, normally
created by \code{\link{makePeptideTable}}}

\item{aggregate.fun}{Function to aggregate peptides into a protein. See
"Details" for more details.}

\item{...}{Additional parameters to pass to \code{aggregate.fun}}

\item{min.peptides}{Minimum number of peptides per protein}

\item{ncores}{Number of cores for parallel processing}
}
\value{
A \code{proteusData} object containing protein intensities and
  metadata.
}
\description{
\code{makeProteinTable} creates a protein table from the peptide table.
Protein intensities or ratios are aggregated from peptide data. An external
function is used to aggregate each protein.
}
\details{
There are three protein aggregation functions provided in this package:
\code{\link{aggregateHifly}}, \code{\link{aggregateMedian}} and
\code{\link{aggregateSum}}. Depending on the needs, the user can provide any
arbitrary function to perform aggregation.

The aggregate function should be in form: \code{function(wp, ...)}. \code{wp}
is a matrix containing measurements from all peptides for the particular
protein. Rows are peptides, columns are samples. \code{...} are additional
parameters for the function that are passed from \code{makeProteinTable}. The
aggregate function should return a vector of values for each sample. That is
the length of the vector should be the same as the number of columns in
\code{wp}.
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)
prodat <- makeProteinTable(pepdat.clean, ncores=2)

}
