% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\docType{data}
\name{measureColumns}
\alias{measureColumns}
\title{Measure columns}
\format{An object of class \code{list} of length 1.}
\usage{
measureColumns
}
\description{
\code{measureColumns} contains measurement columns from evidence file. In
case of label-free data, there is only one column: Intensity.
}
\examples{
str(measureColumns)

}
\keyword{datasets}
