% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize proteus data}
\usage{
normalizeData(pdat, norm.fun = normalizeMedian)
}
\arguments{
\item{pdat}{A \code{proteusData} object with peptide or protein intensities.}

\item{norm.fun}{A normalizing function.}
}
\value{
A \code{proteusData} object with normalized intensities.
}
\description{
\code{normalizeData} normalized the intensity table in a \code{proteusData}
object, using the provided normalizig function.
}
\examples{
protab.norm <- normalizeData(protab, norm.fun=normalizeMedian)

}
