% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{normalizeMedian}
\alias{normalizeMedian}
\title{Normalize intensity table to medians}
\usage{
normalizeMedian(tab)
}
\arguments{
\item{tab}{Data frame with raw intensities. Normally, this is a \code{tab} field in
the \code{proteusData} object (see examples below).}
}
\value{
Normalized intensity table.
}
\description{
\code{normalizeMedian} normalizes intensity table to equalize medians of each sample (column).
}
\examples{
tab <- normalizeMedian(pepdat$tab)

}
