% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{normalizeTable}
\alias{normalizeTable}
\title{Normalize intensity table}
\usage{
normalizeTable(tab, normfac)
}
\arguments{
\item{tab}{Peptab table.}

\item{normfac}{A vector of normalizing factors.}
}
\description{
Normalize intensity table with normalizing factors. Attributes are retained.
}
