% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{normalizeTMT}
\alias{normalizeTMT}
\title{Normalize protein/peptide data using CONSTANd normalization}
\usage{
normalizeTMT(pdat, max.iter = 50, eps = 1e-05)
}
\arguments{
\item{pdat}{A \code{proteusData} object with peptide or protein intensities.}

\item{max.iter}{Maximum number of iterations for the RAS procedure.}

\item{eps}{Convergence limit for the RAS procedure.}
}
\value{
A \code{proteusData} with normalized data.
}
\description{
Normalize protein/peptide data using CONSTANd normalization
}
\details{
\code{normalizeTMT} implements CONSTANd algorithm from Maes et al. (2016)
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4974351/pdf/zjw2779.pdf}.
It normalizes TMT table, for each experiment separately. After normalization
each row shows the precentage of total row intensity and each column is
normalized to their mean value. Hence, both row and column means are equal
1/n, where n is the number of columns (reporters).
}
