% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{parameterString}
\alias{parameterString}
\title{Helper function, not exported}
\usage{
parameterString(...)
}
\arguments{
\item{...}{Any parameters}
}
\value{
A string with list of parameters and their values
}
\description{
Helper function, not exported
}
