% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{Plot correlation matrix}
\usage{
plotCorrelationMatrix(pdat)
}
\arguments{
\item{pdat}{Peptide/protein data object}
}
\description{
Plot correlation matrix for peptide or protein data.
}
