% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotCount}
\alias{plotCount}
\title{Plot peptide or protein count per sample}
\usage{
plotCount(pdat, x.text.size = 10, palette = cbPalette)
}
\arguments{
\item{pdat}{A \code{proteusData} object.}

\item{x.text.size}{Size of text on the x-axis.}

\item{palette}{Palette of colours}
}
\value{
A plot of the number of peptides/proteins detected in each sample.
}
\description{
\code{plotCount} makes a plot of peptide/protein count per sample.
}
\examples{
plotCount(xppepdat)

}
