% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotDetectionSimilarity}
\alias{plotDetectionSimilarity}
\title{Detection Jaccard similarity}
\usage{
plotDetectionSimilarity(pdat, bin.size = 0.01, text.size = 12,
  plot.grid = TRUE, hist.colour = "grey30")
}
\arguments{
\item{pdat}{A \code{proteusData} object with peptides (or proteins).}

\item{bin.size}{Bin size for the histogram.}

\item{text.size}{Text size.}

\item{plot.grid}{Logical to plot grid.}

\item{hist.colour}{Colour of the histogram.}
}
\description{
\code{plotDetectionSimilarit} plots a distribution of (peptide) detection
similarities between samples. Jaccard similarity between two sets is defined
as the size of the intersection divided by the size of the union. This plot
can be used to assess quality of data.
}
\examples{
plotDetectionSimilarity(pepdat)

}
