% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotDistanceMatrix}
\alias{plotDistanceMatrix}
\title{Plot distance matrix}
\usage{
plotDistanceMatrix(pdat, distance = c("correlation"), text.size = 10)
}
\arguments{
\item{pdat}{Peptide or protein \code{proteusData} object.}

\item{distance}{A method to calculate distance.}

\item{text.size}{Text size on axes}
}
\description{
\code{plotDistanceMatrix} plots a distance matrix for peptide or protein
data.
}
\details{
Computes a distance matrix and plots it as a shade-plot. The default distance
is Pearson's correlation coefficient. Other methods will be introduced later.
}
\examples{
plotDistanceMatrix(xppepdat)

}
