% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotFID}
\alias{plotFID}
\title{Fold-change intensity diagram}
\usage{
plotFID(pdat, pair = NULL, bins = 80, marginal.histograms = FALSE,
  xmin = NULL, xmax = NULL, ymax = NULL, text.size = 12,
  point.size = 1, show.legend = TRUE, plot.grid = TRUE, binhex = TRUE,
  transform.fun = log10)
}
\arguments{
\item{pdat}{Protein \code{proteusData} object.}

\item{pair}{A two-element vector containing the pair of conditions to use.
Can be skipped if there are only two conditions.}

\item{bins}{Number of bins for binhex.}

\item{marginal.histograms}{A logical to add marginal histograms.}

\item{xmin}{Lower limit on x-axis.}

\item{xmax}{Upper limit on x-axis.}

\item{ymax}{Upper limit on y-axis. If used, the lower limit is -ymax.}

\item{text.size}{Text size.}

\item{point.size}{Size of points in the plot.}

\item{show.legend}{Logical to show legend (colour key).}

\item{plot.grid}{Logical to plot a grid.}

\item{binhex}{Logical. If TRUE, a hexagonal densit plot is made, otherwise it
is a simple point plot.}

\item{transform.fun}{A function to transform data before plotting.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotFID} makes a log10 fold change versus log10 sum intensity plot,
usually known as MA plot.
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)
prodat.med <- normalizeData(prodat)
plotFID(prodat.med)

}
