% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{plotFID_live}
\alias{plotFID_live}
\title{Fold-change intensity diagram in live Shiny session}
\usage{
plotFID_live(pdat, res, max_points = 100)
}
\arguments{
\item{pdat}{Protein \code{proteusData} object.}

\item{res}{Result table from  \code{\link{limmaDE}}.}

\item{max_points}{Maximum number of points that can be selected.}
}
\value{
A \code{shiny} session hosted locally.
}
\description{
\code{plotFID_live} makes a log10 fold change versus log10 sum intensity plot,
usually known as MA plot.
}
\examples{
library(shiny)
library(proteusLabelFree)
data(proteusLabelFree)
prodat.med <- normalizeData(prodat)
res <- limmaDE(prodat.med,sig.level = 0.05)
\dontrun{
plotFID_live(prodat.med, res)
}

}
