% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotIntensities}
\alias{plotIntensities}
\title{Plot protein/peptide intensities}
\usage{
plotIntensities(
  pdat,
  id = NULL,
  log = FALSE,
  ymin = as.numeric(NA),
  ymax = as.numeric(NA),
  text.size = 12,
  point.size = 3,
  title = NULL
)
}
\arguments{
\item{pdat}{A \code{proteusData} object.}

\item{id}{Protein name, peptide sequence or a vector with these.}

\item{log}{Logical. If set TRUE a logarithm of intensity is plotted.}

\item{ymin}{Lower bound for y-axis}

\item{ymax}{Upper bound for y-axis}

\item{text.size}{Text size}

\item{point.size}{Point size}

\item{title}{Title of the plot (defaults to protein name)}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotIntensities} makes a plot with peptide/protein intensity as a
function of the condition and replicate. When multiple IDs are entered, the
mean and standard error is plotted.
}
\examples{
library(proteusLabelFree)
data(proteusLabelFree)
prodat.med <- normalizeData(prodat)
plotIntensities(prodat.med, id='sp|P26263|PDC6_YEAST', log=TRUE)

}
