% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotIntensityDistributions}
\alias{plotIntensityDistributions}
\title{Plot distribution of intensities/ratios for each sample}
\usage{
plotIntensityDistributions(pdat, title = "Intensity distributions",
  x.text.size = 7, x.text.angle = 90, ymin = as.numeric(NA),
  ymax = as.numeric(NA), logbase = 10, fill = NULL, colour = NULL,
  vline = FALSE)
}
\arguments{
\item{pdat}{A \code{proteusData} object with peptide/protein intensities.}

\item{title}{Title of the plot.}

\item{x.text.size}{Text size in x-axis}

\item{x.text.angle}{Text angle in x-axis}

\item{ymin}{Lower bound on y-axis}

\item{ymax}{Upper bound on y-axis}

\item{logbase}{Base of the logarith which will be applied to data}

\item{fill}{A metadata column to use for the fill of boxes}

\item{colour}{A metadata column to use for the outline colour of boxes}

\item{vline}{Logical, if true a horizontal line at zero is added}
}
\description{
\code{plotIntensityDistributions} makes a boxplot with intensity/ratio distribution for each sample.
}
\examples{
plotIntensityDistributions(prodat)
plotIntensityDistributions(normalizeMedian(prodat))
}
