% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotMA}
\alias{plotMA}
\title{MA plot}
\usage{
plotMA(pdat, pair = NULL, pvalue = NULL, marginal.histograms = FALSE,
  classic = FALSE, xmin = NULL, xmax = NULL, ymax = NULL)
}
\arguments{
\item{pdat}{Protein data structure.}

\item{pair}{A two-element vector containing the pair of conditions to use. Can be skipped if there are only two conditions.}
}
\description{
log fold change versus log sum plot.
}
