% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotMV}
\alias{plotMV}
\title{Plot mean-variance relationship}
\usage{
plotMV(pdat, with.loess = FALSE, bins = 80, xmin = 5, xmax = 10,
  ymin = 7, ymax = 20)
}
\arguments{
\item{pdat}{Peptide or protein \code{proteusData} object.}

\item{with.loess}{Logical. If true, a loess line will be added.}

\item{bins}{Number of bins for binhex}

\item{xmin}{Lower limit on x-axis}

\item{xmax}{Upper limit on x-axis}

\item{ymin}{Lower limit on y-axis}

\item{ymax}{Upper limit on y-axis}
}
\description{
\code{plotMV} makes a plot with variance of log-intensity vs mean of log-intensity.
}
\examples{
plotMV(prodat, with.loess=TRUE)

}
