% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA}
\usage{
plotPCA(
  pdat,
  with.legend = TRUE,
  point.size = 1.2,
  text.size = 10,
  label.size = 3,
  legend.text.size = 7,
  palette = cbPalette
)
}
\arguments{
\item{pdat}{Peptide or protein \code{proteusData} object.}

\item{with.legend}{Logical, whether to plot a legend.}

\item{point.size}{Size of the point in the plot.}

\item{text.size}{Size of the axis text.}

\item{label.size}{Size of the text labels identifying sapmles.}

\item{legend.text.size}{Size of the legend text.}

\item{palette}{Palette of colours}
}
\value{
Creates a plot
}
\description{
\code{plotPCA} creates a principal analysis plot using first two components.
}
\examples{
library(proteusLabelFree)
data(proteusLabelFree)
plotPCA(pepdat)
}
