% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotPdist}
\alias{plotPdist}
\title{Plot p-value distribution}
\usage{
plotPdist(res, bin.size = 0.02, text.size = 12, plot.grid = TRUE)
}
\arguments{
\item{res}{Result table from limmaTable}

\item{text.size}{Text size}

\item{plot.grid}{Logical to plot grid}
}
\description{
Plot distribution of raw p-value, obtained by limmaDE.
}
