% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotPeptideCount}
\alias{plotPeptideCount}
\title{Plot peptide count per sample}
\usage{
plotPeptideCount(pdat, x.text.size = 7)
}
\arguments{
\item{pdat}{Peptide data object}

\item{x.text.size}{Size of text on the x-axis}
}
\value{
A plot of the number of peptides detected in each sample
}
\description{
Plot peptide count per sample
}
