% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotProtPeptides}
\alias{plotProtPeptides}
\title{Protein-peptide plot}
\usage{
plotProtPeptides(pepdat, protein, prodat = NULL, palette = "Accent")
}
\arguments{
\item{pepdat}{Peptide \code{proteusData} object.}

\item{protein}{Protein name.}

\item{prodat}{(optional) protein \code{proteusData} object.}

\item{palette}{Palette of colours from \code{ColorBrewer}}
}
\description{
\code{plotProtPeptides} creates a plot consisting of two panels. The top
panel shows peptide log intensity. Each box represents one peptide, peptide
numbering follows alphabetical sequence order. The bottom panel shows sample
intensity. Each box represents one sample. White boxes show derived protein
intensities (if \code{prodat} is provided).
}
\examples{
plotProtPeptides(xppepdat, "sp|P16522|CDC23_YEAST", prodat = xpprodat)

}
