% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotSampleDistributions}
\alias{plotSampleDistributions}
\title{Plot distribution of intensities/ratios for each sample}
\usage{
plotSampleDistributions(pdat, title = "", method = "dist",
  x.text.size = 7, x.text.angle = 90, vmin = as.numeric(NA),
  vmax = as.numeric(NA), logbase = 10, fill = NULL, colour = NULL,
  hline = FALSE)
}
\arguments{
\item{pdat}{A \code{proteusData} object with peptide/protein intensities.}

\item{title}{Title of the plot.}

\item{method}{"box" or "dist"}

\item{x.text.size}{Text size in x-axis}

\item{x.text.angle}{Text angle in x-axis}

\item{vmin}{Lower bound on log intensity/ratio}

\item{vmax}{Upper bound on log intenstiy/ratio}

\item{logbase}{Base of the logarithm which will be applied to data}

\item{fill}{A metadata column to use for the fill of boxes}

\item{colour}{A metadata column to use for the outline colour of boxes}

\item{hline}{Logical, if true a horizontal line at zero is added}
}
\description{
\code{plotSampleDistributions} makes a boxplot with intensity/ratio distribution for each sample.
}
\examples{
plotSampleDistributions(prodat)
plotSampleDistributions(normalizeMedian(prodat))
}
