% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{plotVolcano_live}
\alias{plotVolcano_live}
\title{Volcano plot in live Shiny session}
\usage{
plotVolcano_live(pdat, res, max_points = 100)
}
\arguments{
\item{pdat}{Protein \code{proteusData} object.}

\item{res}{Result table from  \code{\link{limmaDE}}.}

\item{max_points}{Maximum number of points that can be selected.}
}
\value{
A \code{shiny} session hosted locally.
}
\description{
\code{plotVolcano_live} makes a volcano plot from limma results.
}
\examples{
library(shiny)
library(proteusLabelFree)
data(proteusLabelFree)
prodat.med <- normalizeData(prodat)
res <- limmaDE(prodat.med)
\dontrun{
plotVolcano_live(prodat.med, res)
}

}
