% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\docType{data}
\name{proteinColumns}
\alias{proteinColumns}
\title{Protein columns}
\format{An object of class \code{list} of length 3.}
\usage{
proteinColumns
}
\description{
To be used as the default \code{data.cols} parameter in \code{\link{readProteinGroups}}.
}
\examples{
str(proteinColumns)

}
\keyword{datasets}
