% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{proteinInfo}
\alias{proteinInfo}
\title{proteinInfo}
\usage{
proteinInfo(tab, input, pdat, max_points)
}
\arguments{
\item{tab}{Table used to create a plot}

\item{input}{Input variable from shiny server}

\item{pdat}{\code{proteusData} object with data and annotations}

\item{max_points}{Maximum number of points to select}
}
\value{
Rendered table with protein(s) annotations
}
\description{
Create a rendered table with protein annotations.
}
