% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{proteusData}
\alias{proteusData}
\title{\code{proteusData} constructor}
\usage{
proteusData(tab, metadata, content, pep2prot, peptides, proteins, values,
  npep = NULL, type = "unlabelled", pepseq = "sequence", hifly = NULL,
  min.peptides = NULL, norm.fun = identity, protein.method = NULL)
}
\arguments{
\item{tab}{A matrix with data, rows are peptides/proteins, columns are samples}

\item{metadata}{A data frame with metadata, needs at least two columns: "sample" and "condition"}

\item{content}{Either "peptides" or "proteins"}

\item{pep2prot}{A data frame with two columns: "sequence" and "protein"}

\item{peptides}{A character vector with peptide sequences}

\item{proteins}{A character vector with protein names}

\item{values}{A character vector with names of intensity and/or ratio columns}

\item{npep}{A data frame with number of peptides per protein}

\item{type}{Type of experiment: "unlabelled" or "SILAC"}

\item{pepseq}{Name of the sequence used, either "sequence" or "modseq"}

\item{hifly}{Number of high-flyers}

\item{min.peptides}{Integer, minimum number of peptides to combine into a protein}

\item{norm.fun}{Normalizing function}

\item{protein.method}{Name of the method with which proteins were quantified}
}
\value{
A \code{proteusData} object.
}
\description{
\code{proteusData} constructor
}
