% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{RAS}
\alias{RAS}
\title{RAS procedure for CONSTANd normalization (not exported)}
\usage{
RAS(K, max.iter = 50, eps = 1e-05)
}
\arguments{
\item{K}{Input matrix}

\item{max.iter}{Maximum number of iterations}

\item{eps}{Convergence limit}
}
\value{
Matrix normalized so row and column means equal 1/n (n - number of columns)
}
\description{
RAS procedure for CONSTANd normalization (not exported)
}
