% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{readMaxQuantTable}
\alias{readMaxQuantTable}
\title{Read peptides or proteinGroups file from MaxQuant}
\usage{
readMaxQuantTable(file, meta, id, columns)
}
\arguments{
\item{file}{Input file}

\item{meta}{Metadata}

\item{id}{Name of the column with identifiers (e.g. "Sequence" or "Protein
ID")}

\item{columns}{A vector of intensity/ratio column names}
}
\description{
\code{readMaxQuantTable} reads a MaxQuant's output table (either peptides or
proteinGroups), extracts intensity data and creates a minimal
\code{proteusData} object.
}
