% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{readPeptideFile}
\alias{readPeptideFile}
\title{Read MaxQuant's peptide file}
\usage{
readPeptideFile(file)
}
\arguments{
\item{file}{File name.}
}
\value{
Data frame with selected columns from the evidence file.
}
\description{
\code{readPeptideFile} reads MaxQuant's peptide file. Contaminants and reverse sequences
are filtered out.
}
\examples{
pep.tab <- readPeptideFile("peptides.txt")

}
