% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{readProteinFile}
\alias{readProteinFile}
\title{Read MaxQuant's protein file}
\usage{
readProteinFile(file, meta)
}
\arguments{
\item{file}{File name.}

\item{meta}{Data frame with metadata. As a minimum, it should contain "sample" and "condition" columns.}
}
\value{
A minimal \code{proteusData} object with peptide intensities.
}
\description{
\code{readPeptideFile} reads MaxQuant's protein file and extracts intensity table.
}
\examples{
\dontrun{
protMQ <- readProteinFile("proteinGroups.txt", meta)
}
}
