% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{readProteinGroups}
\alias{readProteinGroups}
\title{Read protein groups file from MaxQuant}
\usage{
readProteinGroups(file, meta, measure.cols = NULL,
  data.cols = proteinColumns)
}
\arguments{
\item{file}{Input file}

\item{meta}{Metadata data frame. As a minimum, it should contain two columns:
\code{sample} with unique sample names and \code{condition} with condition
names.}

\item{measure.cols}{A named list with measure columns to read. These are
columns in the proteinGroup files containing intensity measurements per
sample. Usually they are called \code{Intensity <sample>}. If not provided,
column names will be created from metadata using a template
\code{paste("Intensity", meta$sample)}.}

\item{data.cols}{A named list with columns to read in addition to measure
columns. It should contain three columns named \code{protein}, which will
be used as protein ID and \code{reverse} and \code{contaminant}, which will
be used to filter data. The values of this named list are the actual column
names in the file (including spaces). The default structure
\code{proteinColumns} is provided with the package, though it might need
modifying in case these columns are named differently in the protein groups
file.}
}
\value{
A \code{proteusData} object with protein intensity data.
}
\description{
\code{readProteinGroups} reads a MaxQuant's protein groups file (usually
named \code{proteinGroups.txt}), extracts intensity data and creates a
\code{proteusData} object.
}
\details{
This function allows to read MaxQuant's proteinGroups file directly, without
reading and processing the evidence file. Apart from the proteinGroups file
name, a metadata data frame needs to be provided.
}
\examples{
library(proteusLabelFree)
proteinGroupsFile <- system.file("extdata", "proteinGroups.txt.gz", package="proteusLabelFree")
prot.MQ <- readProteinGroups(proteinGroupsFile, meta)

}
