% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{replicateTable}
\alias{replicateTable}
\title{replicateTable}
\usage{
replicateTable(tab, input, pdat, max_points)
}
\arguments{
\item{tab}{Table used to create a plot}

\item{input}{Input variable from shiny server}

\item{pdat}{\code{proteusData} object with data and annotations}

\item{max_points}{Maximum number of points to select}
}
\value{
Rendered table with replicate intensities.
}
\description{
Create a rendered table with replicate intensities. When multiple points are
selected, the returned table contains mean values. If \code{intensityScale}
in the input is 'log', data are log-10 transformed.
}
