% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{significanceTable}
\alias{significanceTable}
\title{significanceTable}
\usage{
significanceTable(tab, res, input)
}
\arguments{
\item{tab}{Table used to create a plot}

\item{res}{Result from limma}

\item{input}{Input variable from shiny server}
}
\value{
A rendered table with pvalue and adjusted p-value
}
\description{
significanceTable
}
