% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{splitConditions}
\alias{splitConditions}
\title{Split peptide table into conditions}
\usage{
splitConditions(peptab, meta)
}
\arguments{
\item{meta}{Metadata data frame, needs columns 'sample' and 'condition'.}
}
\value{
A list of intensity tables, one per condition.
}
\description{
Split peptide table (containing all samples) into separate tables, one per condition.
Condition information is provided in the meta table.
}
