% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector}
\usage{
standardize(v, trim = 0)
}
\arguments{
\item{v}{Input numeric vector}

\item{trim}{A number of points to be trimmed on each side for mean and sd}
}
\value{
Z-score vector
}
\description{
Calculate Z-scores for a vector, NAs are ignored.
}
