% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{summary.proteusData}
\alias{summary.proteusData}
\title{Summary of \code{proteusData} object}
\usage{
\method{summary}{proteusData}(pdat)
}
\arguments{
\item{pdat}{\code{proteusData} object}
}
\description{
Summary of \code{proteusData} object
}
\examples{
summary(protab)
}
