% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{testNormality}
\alias{testNormality}
\title{Anderson-Darling test for normality}
\usage{
testNormality(w)
}
\arguments{
\item{w}{Intensity table (columns: replicates, rows: peptides/proteins/genes etc.)}
}
\value{
Data frame with mean intensity, p-value and BH-adjusted p-value
}
\description{
Performs a normality test for the intensity table, row by row.
}
