% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{testNormalityConditions}
\alias{testNormalityConditions}
\title{Normality test wrapper}
\usage{
testNormalityConditions(dat, norm = "median")
}
\arguments{
\item{dat}{Intensity data, unnormalized raw data}

\item{norm}{How to normalize intensities (default: "median")}
}
\value{
A data frame with test results
}
\description{
Performs a normality test for the intensity table list (all conditions).
}
