% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{trimVector}
\alias{trimVector}
\title{Trim a vector}
\usage{
trimVector(v, trim)
}
\arguments{
\item{v}{Input vector}

\item{trim}{number of elements to remove on each side}
}
\value{
Trimmed vector
}
\description{
Trim a vector by replacing "trim" top and bottom elements with NA.
}
