% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{tryQuery}
\alias{tryQuery}
\title{Internal function to query UniProt}
\usage{
tryQuery(url, maxtry = 5)
}
\arguments{
\item{url}{Full query URL}

\item{maxtry}{Maximum number of tries}
}
\value{
Query result
}
\description{
Internal function to query UniProt
}
