% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xppepdat.clean}
\alias{xppepdat.clean}
\title{Clean peptide data}
\format{A \code{proteusData} object}
\usage{
data(xppepdat.clean)
}
\description{
Peptide data aggregated from evidence data \code{xpevi}. One "bad" replicate was removed.
}
