% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xppepdat}
\alias{xppepdat}
\title{Peptide data}
\format{A \code{proteusData} object}
\usage{
data(xppepdat)
}
\description{
Peptide data aggregated from evidence data \code{xpevi}. Contains all replicates.
}
