% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{aggregateHifly}
\alias{aggregateHifly}
\title{Protein aggregate function using hifly method}
\usage{
aggregateHifly(wp, hifly = 3)
}
\arguments{
\item{wp}{Intensity table with selection of peptides for this protein}

\item{hifly}{Number of high-fliers}
}
\value{
A data frame row with aggregated protein intensities
}
\description{
Protein aggregate function using hifly method
}
\details{
This function should be used from within
\code{\link{makeProteinTable}}. The "hifly" method is a follows. \enumerate{
\item For a given protein find all corresponding peptides. \item In each
replicate, order intensities from the highest to the lowest. This is done
separetly for each replicate. \item Select n top rows of the ordered table.
\item In each replicate, find the mean of these three rows. This is the
estimated protein intensity. }
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)
prodat <- makeProteinTable(pepdat.clean, aggregate.fun=aggregateHifly, hifly=3)

}
