% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{aggregateMedian}
\alias{aggregateMedian}
\title{Protein aggregate function using median}
\usage{
aggregateMedian(wp)
}
\arguments{
\item{wp}{Intensity table with selection of peptides for this protein}
}
\value{
A data frame row with aggregated protein intensities
}
\description{
Protein aggregate function using median
}
\details{
This function should be used from within
\code{\link{makeProteinTable}}. It aggregates peptides into a protein by
calculating the median for each sample.
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)
prodat <- makeProteinTable(pepdat.clean, aggregate.fun=aggregateMedian)

}
