% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{aggregateSum}
\alias{aggregateSum}
\title{Protein aggregate function using sum}
\usage{
aggregateSum(wp)
}
\arguments{
\item{wp}{Intensity table with selection of peptides for this protein}
}
\value{
A data frame row with aggregated protein intensities
}
\description{
Protein aggregate function using sum
}
\details{
This function should be used from within
\code{\link{makeProteinTable}}. It aggregates peptides into a protein by
calculating the sum for each sample.
}
