% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\docType{data}
\name{allowedUniProtColumns}
\alias{allowedUniProtColumns}
\title{UniProt columns}
\format{An object of class \code{character} of length 27.}
\usage{
allowedUniProtColumns
}
\description{
A character vector containing possible columns to query using \code{\link{fetchFromUniProt}}.
}
\examples{
allowedUniProtColumns

}
\keyword{datasets}
