% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\docType{data}
\name{evidenceColumns}
\alias{evidenceColumns}
\title{Evidence columns}
\format{
An object of class \code{list} of length 9.
}
\usage{
evidenceColumns
}
\description{
\code{evidenceColumns} contains default columns to be read from the evidence
file. The names of list elements are used internally to reference evidence
data.
}
\examples{
str(evidenceColumns)

}
\keyword{datasets}
