% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{fetchFromUniProt}
\alias{fetchFromUniProt}
\title{Fetch annotations from UniProt}
\usage{
fetchFromUniProt(unis, columns = c("genes", "protein names"),
  col.names = NULL, batchsize = 400, verbose = FALSE)
}
\arguments{
\item{unis}{A character vector with UniProt identifiers}

\item{columns}{Data columns requested}

\item{col.names}{How to name data columns in the returned data frame}

\item{batchsize}{Size of batch of proteins in a single query}

\item{verbose}{Logical, if true, query progress will be displayed}
}
\value{
A data frame with protein annotations.
}
\description{
Fetch annotations from UniProt
}
\details{
For a given list of UniProt identifiers this function will bring back
annotations from UniProt servers. What information is downloaded is
controlled by the \code{columns} parameter. By default it fetches gene names
and protein name/description. The full list of available columns is in a
vector \code{allowedUniProtColumns}.

The column names in the returned data frame are the same as in \code{columns}
parameter, unless alternative names are provided in parameter
\code{col.names}. The \code{id} column is added by default.
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)

# Extract UniProt identifiers from protein IDs
unis <- sapply(as.character(prodat$proteins), function(prot) {
 s <- unlist(strsplit(prot, "|", fixed=TRUE))
 s[2]
})

# Fetch first 100 annotations (for a quick example)
anno <- fetchFromUniProt(unis[1:100])
}
