% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{intensityStats}
\alias{intensityStats}
\title{Statistics for an intensity table}
\usage{
intensityStats(pdat)
}
\arguments{
\item{pdat}{Intensity dat tructure}
}
\value{
A data frame with several statistics
}
\description{
Calculate mean, standard deviation, ... for an intensity table (that is peptide or protein,
for all conditions)
}
