% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{intensityStats}
\alias{intensityStats}
\title{Statistics for an intensity table}
\usage{
intensityStats(pdat)
}
\arguments{
\item{pdat}{Peptide or protein \code{proteusData} object.}
}
\value{
A data frame with several statistics.
}
\description{
\code{intensityStats} calculates the mean, variance and number of good data
points for peptide or protein intensities.
}
\examples{
stats <- intensityStats(xpprodat)

}
