% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{limmaDE}
\alias{limmaDE}
\title{Differential expression with limma}
\usage{
limmaDE(pdat, formula = "~condition")
}
\arguments{
\item{pdat}{Protein \code{proteusData} object.}

\item{formula}{A string with a formula for building the linear model.}
}
\value{
limma output from eBays. See limma documentation for more details.
The output from this function can be used with \code{\link{limmaTable}} to create
a table with differential expression results.
}
\description{
\code{limmaDE} is a simple wrapper around limma differential expression. It performs
differential expression on the intensity table.
}
\examples{
ebay <- limmaDE(prodat, formula="~condition + batch + condition:batch")
res <- limmaTable(prodat, ebay)

}
