% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{limmaDE}
\alias{limmaDE}
\title{Simple differential expression with limma}
\usage{
limmaDE(pdat, formula = "~condition", conditions = NULL)
}
\arguments{
\item{pdat}{Protein \code{proteusData} object.}

\item{formula}{A string with a formula for building the linear model.}

\item{conditions}{A character vector with two conditions for differential
expression. Can be omitted if there are only two condition in \code{pdat}.}
}
\value{
A data frame with DE results. "logFC" colum is a log10-fold-change.
}
\description{
\code{limmaDE} is a wrapper around \code{\link{limma}} to performa a differential
expression between a pair of conditions.
}
\details{
Before \code{limma} is called, intensity data are log-transformed
  using base-10 logarithms. Therefore, the column "logFC" in the output data
  frame contains log10 fold change (not log2!). If you need log2-based fold
  change, you need to convert the column: `res$logFC <- res$logFC /
  log10(2)`.
}
\examples{
res <- limmaDE(xpprodat)

}
