% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{limmaDE}
\alias{limmaDE}
\title{Differential expression with limma}
\usage{
limmaDE(pdat, formula = "~condition")
}
\arguments{
\item{pdat}{Protein intensity structure.}

\item{formula}{A string with a formula for building the linear model. The default value is "~condition".}
}
\value{
limma output from eBays. See limma documentation for more details.
}
\description{
Calls limma to perform differential expression.
}
