% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{limmaTable}
\alias{limmaTable}
\title{Create differential expression result}
\usage{
limmaTable(pdat, ebay, column = "condition")
}
\arguments{
\item{pdat}{Protein \code{proteusData} object.}

\item{ebay}{Output from  \code{\link{limmaDE}}.}

\item{column}{Which column should be used to extract data. The default value
is "condition".}
}
\value{
A data frame with DE results. "logFC" colum is a log10-fold-change.
}
\description{
\code{limmaTable} creates a table with differential expression results, using
an object created with \code{\link{limmaDE}}
}
\details{
Before \code{limma} is called, intensity data are log-transformed
  using base-10 logarithms. Therefore, the column "logFC" in the output data
  frame contains log10 fold change (not log2!). If you need log2-based fold
  change, you need to convert the column: `res$logFC <- res$logFC /
  log10(2)`.
}
\examples{
ebay <- limmaDE(prodat, formula="~condition")
res <- limmaTable(prodat, ebay)

}
