% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{makePeptideTable}
\alias{makePeptideTable}
\title{Create peptide table from evidence data}
\usage{
makePeptideTable(evi, meta, id = "sequence", value = "intensity")
}
\arguments{
\item{evi}{Evidence table created with readEvidenceFile.}

\item{meta}{Data frame with metadata. As minimum, it should contain "sample" and "condition".}

\item{id}{A column name to identify peptides. The default is "sequence". Can be "modseq".}

\item{value}{A column name to use for results. The default is "intensity".}
}
\value{
An intensity table structure
}
\description{
Creates a table with columns corresponding to samples (experiments) and rows corresponding to peptides.
Each cell is a sum of all intensities for this sample/peptide in the input evidence data.
}
