% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{makePeptideTable}
\alias{makePeptideTable}
\title{Create peptide table from evidence data}
\usage{
makePeptideTable(evi, meta, pepseq = "sequence", values = "intensity",
  fun.aggregate = sum, experiment.type = "unlabelled")
}
\arguments{
\item{evi}{Evidence table created with readEvidenceFile.}

\item{meta}{Data frame with metadata. As a minimum, it should contain
"sample" and "condition" columns.}

\item{pepseq}{A column name to identify peptides. Can be either "sequence" or
"modseq".}

\item{values}{A vector of names (or one name) of column(s) with intensity/ratio data to be used.}

\item{fun.aggregate}{A function to aggregate pepetides with the same sequence/sample.}
}
\value{
A \code{proteusData} object, containing peptide intensities and
  metadata.
}
\description{
\code{makePeptideTable} creates a table with columns corresponding to samples
(experiments) and rows corresponding to peptides. Each cell is a sum of all
intensities for this sample/peptide in the input evidence data.
}
\examples{
pepdat <- makePeptideTable(evi, meta)

}
