% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{makePeptideTable}
\alias{makePeptideTable}
\title{Create peptide table from evidence data}
\usage{
makePeptideTable(evi, meta, pepseq = "sequence", intensity = "intensity")
}
\arguments{
\item{evi}{Evidence table created with readEvidenceFile.}

\item{meta}{Data frame with metadata. As a minimum, it should contain "sample" and "condition" columns.}

\item{pepseq}{A column name to identify peptides. Can be either "sequence" or "modseq".}

\item{intensity}{A column name to use for results. The default is "intensity".}
}
\value{
A \code{proteusData} object, containing peptide intensities and metadata.
}
\description{
\code{makePeptideTable} creates a table with columns corresponding to samples (experiments) and
rows corresponding to peptides. Each cell is a sum of all intensities for this sample/peptide
in the input evidence data.
}
\examples{
evi <- readEvidenceFile("evidence.txt")
meta <- read.delim("metadata.txt", header=TRUE)
meta$sample <- factor(meta$sample, levels=meta$sample)
pepdat <- makePeptideTable(evi, meta)

}
