% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{makeProteinTable}
\alias{makeProteinTable}
\title{Make protein table}
\usage{
makeProteinTable(evi, peptab, hifly = 3, norm = "median",
  min.peptides = 1)
}
\arguments{
\item{evi}{Evidence table, used only to cross-reference peptides with proteins.}

\item{peptab}{Peptide intensity table.}

\item{hifly}{The number of top peptides (high-flyers) to be used for protein intensity.}

\item{norm}{Normalization to use on peptides before converting into proteins.}

\item{min.peptides}{Minimum number of peptides per protein.}
}
\value{
Protein intensity table.
}
\description{
Create a protein table from the peptide table using high-flyers.
}
