% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{makeProteinTable}
\alias{makeProteinTable}
\title{Make protein table}
\usage{
makeProteinTable(pepdat, hifly = 3, norm = "median", min.peptides = 1,
  verbose = FALSE)
}
\arguments{
\item{pepdat}{Peptide intensity object.}

\item{hifly}{The number of top peptides (high-flyers) to be used for protein intensity.}

\item{norm}{Normalization to use on peptides before converting into proteins.}

\item{min.peptides}{Minimum number of peptides per protein.}
}
\value{
Protein intensity object.
}
\description{
Create a protein table from the peptide table using high-flyers.
}
