% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize proteus data}
\usage{
normalizeData(pdat, norm.fun = normalizeMedian)
}
\arguments{
\item{pdat}{A \code{proteusData} object with peptide or protein intensities.}

\item{norm.fun}{A normalizing function.}
}
\value{
A \code{proteusData} object with normalized intensities.
}
\description{
\code{normalizeData} normalizes the intensity table in a \code{proteusData}
object, using the provided normalizing function.
}
\details{
The normalizing function, specified by \code{norm.fun} needs to
  normalize columns of a numerical matrix. The input is a matrix and the
  output is a normalized matrix. The default value points to
  \code{\link{normalizeMedian}}, Proteus's function normalizing each column
  to its median. Other functions can be used, for example
  \code{normalizeQuantiles} from the \code{limma} package.
}
\examples{
library(proteusLabelFree)
data(proteusLabelFree)
prodat.med <- normalizeData(prodat, norm.fun=normalizeMedian)

}
