% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{normalizeMedian}
\alias{normalizeMedian}
\title{Normalize columns of a matrix to medians}
\usage{
normalizeMedian(tab)
}
\arguments{
\item{tab}{Data frame with raw intensities. Normally, this is a \code{tab}
field in the \code{proteusData} object (see examples below).}
}
\value{
Normalized matrix.
}
\description{
\code{normalizeMedian} normalizes the columns of a matrix to have the same
medians. It should be used with \code{\link{normalizeData}} function.
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)
normtab <- normalizeMedian(prodat$tab)

}
