% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{normalizeTable}
\alias{normalizeTable}
\title{Normalize peptide table}
\usage{
normalizeTable(peptab, normfac)
}
\arguments{
\item{peptab}{Peptab table.}

\item{normfac}{A vector of normalizing factors.}
}
\description{
Normalize peptide table with normalizing factors. Attributes are retained.
}
