% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{normalizeTable}
\alias{normalizeTable}
\title{Normalize intensity table}
\usage{
normalizeTable(tab, normfac)
}
\arguments{
\item{tab}{Data frame with raw intensities. Normally, this is a \code{tab} field in
the \code{proteusData}
object (see examples below).}

\item{normfac}{A vector of normalizing (dividing) factors.}
}
\value{
Normalized intensity table.
}
\description{
\code{normalizeTable} normalizes intensity table with normalizing factors, normally obtained with
\code{\link{normalizingFactors}}.
}
\examples{
norm.fac <- normalizingFactors(pepdat$tab)
tab <- normalizeTable(pepdat$tab, norm.fac)

}
