% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{normalizingFactors}
\alias{normalizingFactors}
\title{Calculate normalizing factors}
\usage{
normalizingFactors(tab, method = "median")
}
\arguments{
\item{tab}{Intensity table for all conditions.}

\item{method}{Method of normalization, only 'median' at the moment.}
}
\value{
A vector with normalizing factors.
}
\description{
Calculate normalizing factors
}
