% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotClustering}
\alias{plotClustering}
\title{Plot clustering dendrogram}
\usage{
plotClustering(pdat)
}
\arguments{
\item{pdat}{Peptide or protein \code{proteusData} object.}
}
\value{
Creates a plot
}
\description{
\code{plotClustering} plots a dendrogram of intensity data, using
hierarchical clustering.
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)
plotClustering(pepdat)

}
