% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{Plot correlation matrix}
\usage{
plotCorrelationMatrix(pdat, text.size = 10, palette = "Blues")
}
\arguments{
\item{pdat}{Peptide or protein \code{proteusData} object.}

\item{text.size}{Text size on axes}

\item{palette}{Palette of colours from \code{ColorBrewer}}
}
\description{
\code{plotCorrelationMatrix} plots a correlation matrix for peptide or
protein data.
}
\examples{
plotCorrelationMatrix(xppepdat)

}
