% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{Plot correlation matrix}
\usage{
plotCorrelationMatrix(pdat)
}
\arguments{
\item{pdat}{Peptide or protein \code{proteusData} object.}
}
\description{
\code{plotCorrelationMatrix} plots a correlation matrix for peptide or protein data.
}
\examples{
plotCorrelationMatrix(pepdat)

}
