% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotIntensities}
\alias{plotIntensities}
\title{Plot protein/peptide intensities}
\usage{
plotIntensities(pdat, id = NULL, log = FALSE, ymin = as.numeric(NA),
  ymax = as.numeric(NA), text.size = 12, point.size = 3, title = NULL)
}
\arguments{
\item{pdat}{A \code{proteusData} object.}

\item{log}{Logical. If set TRUE a logarithm of intensity is plotted.}

\item{ymin}{Lower bound for y-axis}

\item{ymax}{Upper bound for y-axis}

\item{text.size}{Text size}

\item{point.size}{Point size}

\item{title}{Title of the plot (defaults to protein name)}

\item{is}{Protein name, peptide sequence or a vector with these.}
}
\description{
\code{plotIntensities} makes a plot with peptide/protein intensity as a function of the
condition and replicate. When multiple IDs are entered, the mean and
standard error is plotted.
}
\examples{
plotIntensities(xpprodat, id="sp|P16522|CDC23_YEAST", log=TRUE, title="MNT2")

}
