% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotMA}
\alias{plotMA}
\title{MA plot}
\usage{
plotMA(pdat, pair = NULL, pvalue = NULL, bins = 80,
  marginal.histograms = FALSE, classic = FALSE, xmin = NULL,
  xmax = NULL, ymax = NULL, text.size = 12, show.legend = TRUE,
  plot.grid = TRUE)
}
\arguments{
\item{pdat}{Protein data structure.}

\item{pair}{A two-element vector containing the pair of conditions to use. Can be skipped if there are only two conditions.}

\item{pvalue}{A vector with corresponding p-values for an interactive plotly plot}

\item{bins}{Number of bins for binhex}

\item{marginal.histograms}{A logical to add marginal histograms}

\item{xmin}{Lower limit on x-axis}

\item{xmax}{Upper limit on x-axis}

\item{ymax}{Upper limit on y-axis. If used, the lower limit is -ymax}

\item{text.size}{Text size}

\item{show.legend}{Logical to show legend (colour key)}

\item{plot.grid}{Logical to plot grid}
}
\description{
log fold change versus log sum plot.
}
