% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotMV}
\alias{plotMV}
\title{Plot mean-variance relationship}
\usage{
plotMV(pdat, with.loess = FALSE, xmin = 5, xmax = 10, ymin = 7,
  ymax = 20)
}
\arguments{
\item{pdat}{Intensity data structure.}

\item{with.loess}{Logical. If true, a loess line will be added.}
}
\description{
Plot variance of log-intensity vs mean of log-intensity.
}
