% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotPdist}
\alias{plotPdist}
\title{Plot p-value distribution}
\usage{
plotPdist(res, bin.size = 0.02, text.size = 12, plot.grid = TRUE)
}
\arguments{
\item{res}{Output table from \code{\link{limmaDE}}.}

\item{bin.size}{Bin size for the histogram.}

\item{text.size}{Text size.}

\item{plot.grid}{Logical to plot grid.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotPdist} makes a plot with distribution of raw p-values, obtained by
\code{\link{limmaDE}}.
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)
prodat.med <- normalizeData(prodat)
res <- limmaDE(prodat.med)
plotPdist(res)

}
