% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotPeptideCount}
\alias{plotPeptideCount}
\title{Plot peptide count per sample}
\usage{
plotPeptideCount(peptab)
}
\arguments{
\item{peptab}{Peptide table (with metadata attribute)}
}
\value{
A plot of the number of peptides detected in each sample
}
\description{
Plot peptide count per sample
}
