% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotPeptideCount}
\alias{plotPeptideCount}
\title{Plot peptide count per sample}
\usage{
plotPeptideCount(pdat, x.text.size = 10)
}
\arguments{
\item{pdat}{Peptide \code{proteusData} object.}

\item{x.text.size}{Size of text on the x-axis.}
}
\value{
A plot of the number of peptides detected in each sample.
}
\description{
\code{plotPeptideCount} makes a plot of peptide count per sample.
}
\examples{
plotPeptideCount(pepdat)

}
