% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotProteins}
\alias{plotProteins}
\title{Plot protein(s)}
\usage{
plotProteins(pdat, protein = protein, log = FALSE, ymin = as.numeric(NA),
  ymax = as.numeric(NA))
}
\arguments{
\item{pdat}{Protein intensity structure.}

\item{protein}{Protein name (string) or a vector with protein names.}

\item{log}{Logical. If set TRUE a logarithm of intensity is plotted.}

\item{ymin}{Lower bound for y-axis}

\item{ymax}{Upper bound for y-axis}
}
\description{
Plot protein intensity as a function of the condition and replicate. When
multiple proteins are entered, the mean and standard error is plotted.
}
