% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotProteins}
\alias{plotProteins}
\title{Plot protein(s)}
\usage{
plotProteins(pdat, protein = protein, log = FALSE, ymin = as.numeric(NA),
  ymax = as.numeric(NA), text.size = 12, point.size = 3, title = NULL)
}
\arguments{
\item{pdat}{Protein \code{proteusData} object.}

\item{protein}{Protein name (string) or a vector with protein names.}

\item{log}{Logical. If set TRUE a logarithm of intensity is plotted.}

\item{ymin}{Lower bound for y-axis}

\item{ymax}{Upper bound for y-axis}

\item{text.size}{Text size}

\item{point.size}{Point size}

\item{title}{Title of the plot (defaults to protein name)}
}
\description{
\code{plotProteins} makes a plot with protein intensity as a function of the
condition and replicate. When multiple proteins are entered, the mean and
standard error is plotted.
}
\examples{
plotProteins(prodat, "sp|P16522|CDC23_YEAST", log=TRUE, title="MNT2")

}
