% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotSampleDistributions}
\alias{plotSampleDistributions}
\title{Plot distribution of intensities/ratios for each sample}
\usage{
plotSampleDistributions(pdat, title = "", method = c("violin", "dist",
  "box"), x.text.size = 7, n.grid.rows = 3, hist.bins = 100,
  x.text.angle = 90, vmin = as.numeric(NA), vmax = as.numeric(NA),
  log.scale = TRUE, log.base = 10, palette = cbPalette, fill = NULL,
  colour = NULL, hline = FALSE)
}
\arguments{
\item{pdat}{A \code{proteusData} object with peptide/protein intensities.}

\item{title}{Title of the plot.}

\item{method}{"box", "violin" or "dist"}

\item{x.text.size}{Text size in value axis}

\item{n.grid.rows}{Number of rows in the grid of facets}

\item{hist.bins}{Number of bins in histograms}

\item{x.text.angle}{Text angle in value axis}

\item{vmin}{Lower bound on log value}

\item{vmax}{Upper bound on log value}

\item{log.scale}{Logical, to plot in logarithmic scale}

\item{log.base}{Base of the logarithm which will be applied to data}

\item{palette}{Palette of colours}

\item{fill}{A metadata column to use for the fill of boxes}

\item{colour}{A metadata column to use for the outline colour of boxes}

\item{hline}{Logical, if true a horizontal line at zero is added}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotSampleDistributions} makes a boxplot with intensity/ratio
distribution for each sample.
}
\examples{
library(proteusUnlabelled)
data(proteusUnlabelled)
plotSampleDistributions(prodat)
plotSampleDistributions(normalizeData(prodat))

}
