% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(res, bins = 80, xmax = NULL, ymax = NULL, text.size = 12,
  show.legend = TRUE, plot.grid = TRUE)
}
\arguments{
\item{res}{Result table from limmaTable}

\item{bins}{Number of bins for binhex}

\item{xmax}{Upper limit on x-axis. If used, the lower limit is -xmax}

\item{ymax}{Upper limit on y-axis. If used, the lower limit is -ymax}

\item{text.size}{Text size}

\item{show.legend}{Logical to show legend (colour key)}

\item{plot.grid}{Logical to plot grid}
}
\description{
Volcano plot from limma results.
}
