% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{proteusData}
\alias{proteusData}
\title{\code{proteusData} constructor}
\usage{
proteusData(tab, metadata, content, pep2prot, peptides, proteins, measures,
  npep = NULL, type = "label-free", sequence.col = "sequence",
  protein.col = "protein", peptide.aggregate.fun = NULL,
  peptide.aggregate.parameters = NULL, protein.aggregate.fun = NULL,
  protein.aggregate.parameters = NULL, min.peptides = NULL,
  norm.fun = identity)
}
\arguments{
\item{tab}{A matrix with data, rows are peptides/proteins, columns are
samples}

\item{metadata}{A data frame with metadata, needs at least two columns:
"sample" and "condition"}

\item{content}{Either "peptides" or "proteins"}

\item{pep2prot}{A data frame with two columns: "sequence" and "protein". Can
use "other" for non-proteus data.}

\item{peptides}{A character vector with peptide sequences}

\item{proteins}{A character vector with protein names}

\item{measures}{A character vector with names of intensity and/or ratio
columns}

\item{npep}{A data frame with number of peptides per protein}

\item{type}{Type of experiment: "label-free" or "SILAC"}

\item{sequence.col}{Name of the sequence used, either "sequence" or "modified_sequence"}

\item{protein.col}{Name of the protein column used, either "protein" or "protein_group"}

\item{peptide.aggregate.fun}{Function used to aggregate peptides}

\item{peptide.aggregate.parameters}{Additional parameters passed to peptide
aggregate function}

\item{protein.aggregate.fun}{Function used to aggregate proteins}

\item{protein.aggregate.parameters}{Additional parameters passed to protein
aggregate function}

\item{min.peptides}{Integer, minimum number of peptides to combine into a
protein}

\item{norm.fun}{Normalizing function}
}
\value{
A \code{proteusData} object.
}
\description{
\code{proteusData} constructor
}
