% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{readColumnNames}
\alias{readColumnNames}
\title{Read column names from a tab-delimited text file}
\usage{
readColumnNames(file)
}
\arguments{
\item{file}{File name.}
}
\value{
A vector with column names.
}
\description{
Read column names from a tab-delimited text file
}
\examples{
evidenceFile <- system.file("extdata", "evidence.txt.gz", package="proteusLabelFree")
evidence.columns <- readColumnNames(evidenceFile)
evidence.columns
}
