% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{readEvidenceFile}
\alias{readEvidenceFile}
\title{Read MaxQuant's evidence file}
\usage{
readEvidenceFile(file, measure.cols = measureColumns,
  data.cols = evidenceColumns)
}
\arguments{
\item{file}{File name.}

\item{measure.cols}{Named list with measure columns to read.}

\item{data.columns}{Named list with other columns to read (in addition to
measure columns).}
}
\value{
Data frame with selected columns from the evidence file.
}
\description{
\code{readEvidenceFile} reads MaxQuant's evidence file. Contaminants and
reverse sequences are filtered out.
}
\details{
There are two parameters controlling which columns are read from the evidence
file. Parameter \code{measure.cols} selects columns with measurements: these
are intensities (unlabelled, TMT) or ratios (Silac). In the simplest case of
unlabelled data, there is only one measure column: "Intensity". Parameter
\code{data.columns} selects all other columns read from the evidence file.
There are two default lists, supplied with the package, appropriate for an
unlabelled experiment, \code{measureColumns} and \code{evidenceColumns}.
}
\examples{
\dontrun{
evidenceFile <- system.file("extdata", "evidence.txt.gz", package="proteus")
xpevi <- readEvidenceFile(evidenceFile)
}

}
