% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{readEvidenceFile}
\alias{readEvidenceFile}
\title{Read MaxQuant's evidence file}
\usage{
readEvidenceFile(file, columns = evidenceColumns)
}
\arguments{
\item{file}{File name.}

\item{columns}{Named list with columns to read.}
}
\value{
Data frame with selected columns from the evidence file.
}
\description{
Works only with unlabelled data, that is, the evidence file with one Intensity column.
Contaminants and reverse sequences are filtered out.
}
