% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{readMaxQuantTable}
\alias{readMaxQuantTable}
\title{Read MaxQuant's table}
\usage{
readMaxQuantTable(file, content, col.id, meta)
}
\description{
\code{readMaxQuantTable} reads a MaxQuant's output table (either peptides or proteinGroups),
 extracts intensity data and creates a minimal \code{proteusData} object.
}
\details{
@param file Input file
 @param content Either "peptide" or "protein"
 @param col.id Name of the column with identifiers (e.g. "Sequence")
 @param meta Metadata
}
