% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{readPeptideFile}
\alias{readPeptideFile}
\title{Read MaxQuant's peptide file}
\usage{
readPeptideFile(file)
}
\arguments{
\item{file}{File name.}
}
\value{
Data frame with selected columns from the evidence file.
}
\description{
Read MaxQuant's peptide file
}
