% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{summary.proteusData}
\alias{summary.proteusData}
\title{Summary of \code{proteusData} object}
\usage{
\method{summary}{proteusData}(object, ...)
}
\arguments{
\item{object}{\code{proteusData} object.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Text output with object summary
}
\description{
Summary of \code{proteusData} object
}
\examples{
library(proteusLabelFree)
data(proteusLabelFree)
summary(prodat)

}
