% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{testNormalityConditions}
\alias{testNormalityConditions}
\title{Normality test wrapper}
\usage{
testNormalityConditions(int)
}
\arguments{
\item{int}{Intensity table list (created with splitConditions)}
}
\value{
A list of data frames with mean, p-value and adjusted p-value
}
\description{
Performs a normality test for the intensity table list (all conditions).
}
