% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{testNormalityConditions}
\alias{testNormalityConditions}
\title{Normality test wrapper}
\usage{
testNormalityConditions(intab, norm = "median")
}
\arguments{
\item{intab}{Intensity table, unnormalized raw data}

\item{norm}{How to normalize intensities (default: "median")}
}
\value{
A list of data frames with mean, p-value and adjusted p-value
}
\description{
Performs a normality test for the intensity table list (all conditions).
}
